Move the gate-level simulator backend to a separate file
