+++ src/qutip_qip/circuit/__init__.py
@@ -1,3 +1,4 @@
 """Circuit representation and simulation at the gate level."""
 from .circuit import *
 from .circuit_latex import *
+from .circuitsimulator import *
